% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3css_animate.R
\name{w3_animate_top}
\alias{w3_animate_top}
\alias{w3_animate_bottom}
\alias{w3_animate_left}
\alias{w3_animate_right}
\alias{w3_animate_opacity}
\alias{w3_animate_zoom}
\alias{w3_animate_fading}
\alias{w3_spin}
\title{W3.CSS Animations}
\usage{
w3_animate_top()

w3_animate_bottom()

w3_animate_left()

w3_animate_right()

w3_animate_opacity()

w3_animate_zoom()

w3_animate_fading()

w3_spin()
}
\value{
A W3 CSS class.
}
\description{
Implementation of W3.CSS Animations, as described in
\url{https://www.w3schools.com/w3css/w3css_animate.asp}
}
\examples{
w3_animate_top()
}
