% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3css_input.R
\name{w3_radioButton}
\alias{w3_radioButton}
\alias{w3_selectInput}
\alias{w3_textInput}
\alias{w3_passwordInput}
\alias{w3_dateInput}
\alias{w3_editableListInput}
\title{W3.CSS Input}
\usage{
w3_radioButton(inputId, label, choices)

w3_selectInput(inputId, label, choices)

w3_textInput(inputId, label, placeholder = "")

w3_passwordInput(inputId, label, placeholder = "")

w3_dateInput(inputId, label, value = Sys.Date(), min = "", max = "")

w3_editableListInput(inputId, label, choices, ul_class = "", li_class = "")
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{choices}{List of values to select from (if elements of the list are
named then that name rather than the value is displayed to the user). If
this argument is provided, then \code{choiceNames} and \code{choiceValues} must not
be provided, and vice-versa. The values should be strings; other types
(such as logicals and numbers) will be coerced to strings.}

\item{placeholder}{A character string giving the user a hint as to what can
be entered into the control. Internet Explorer 8 and 9 do not support this
option.}

\item{value}{Initial value.}

\item{min}{The minimum allowed date. Either a Date object, or a string in
\code{yyyy-mm-dd} format.}

\item{max}{The maximum allowed date. Either a Date object, or a string in
\code{yyyy-mm-dd} format.}

\item{ul_class, li_class}{CSS class for the internal elements.}
}
\value{
Input UI elements
}
\description{
Implementation of W3.CSS Input, as described in
\url{https://www.w3schools.com/w3css/w3css_input.asp}
}
\examples{

w3_actionButton(
"w3_actionButton",
"Show Modal"
)

}
