% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3css_modal.R
\name{w3_modal}
\alias{w3_modal}
\alias{w3_show_modal}
\alias{w3_hide_modal}
\title{W3.CSS Modal}
\usage{
w3_modal(
  inputId,
  title = "",
  body = "",
  footer = "",
  color = "",
  display = "none"
)

w3_show_modal(inputId, session = shiny::getDefaultReactiveDomain())

w3_hide_modal(inputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{title, body, footer}{Content to add in the modal}

\item{color}{W3 CSS Color}

\item{display}{Display style of the modal (either 'none' or 'block')}

\item{session}{The \code{session} object passed to function given to
\code{shinyServer}. Default is \code{getDefaultReactiveDomain()}.}
}
\value{
a modal UI, or used for side effects.
}
\description{
Implementation of W3.CSS Modal, as described in
\url{https://www.w3schools.com/w3css/w3css_modal.asp}
}
\examples{
w3_modal("one modal")
}
