% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3css_responsive.R
\name{w3_half}
\alias{w3_half}
\alias{w3_third}
\alias{w3_twothird}
\alias{w3_quarter}
\alias{w3_threequarter}
\alias{w3_rest}
\alias{w3_col}
\title{W3.CSS Built-In Responsiveness}
\usage{
w3_half(...)

w3_third(...)

w3_twothird(...)

w3_quarter(...)

w3_threequarter(...)

w3_rest(...)

w3_col(...)
}
\arguments{
\item{...}{UI elements to include}
}
\value{
a UI element.
}
\description{
Implementation of W3.CSS Built-In Responsiveness, as described in
\url{https://www.w3schools.com/w3css/w3css_responsive.asp}
}
\examples{
w3_half("hey there")
}
