% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todo_w3css_text.R
\name{w3_left_align}
\alias{w3_left_align}
\alias{w3_right_align}
\alias{w3_center}
\alias{w3_wide}
\alias{w3_opacity}
\title{W3.CSS Text}
\usage{
w3_left_align()

w3_right_align()

w3_center()

w3_wide()

w3_opacity()
}
\value{
A CSS Class
}
\description{
Implementation of W3.CSS Text, as described in
\url{https://www.w3schools.com/w3css/w3css_text.asp}
}
\examples{
w3_left_align()
}
