% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3css_buttons.R
\name{w3_actionButton}
\alias{w3_actionButton}
\title{W3.CSS Buttons}
\usage{
w3_actionButton(
  inputId,
  label,
  type = c("btn", "button", "bar", "block", "circle", "ripple"),
  ...
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{The contents of the button or link--usually a text label, but
you could also use any other HTML, like an image.}

\item{type}{Class to apply to the button}

\item{...}{Named attributes to be applied to the button or link.}
}
\value{
a UI element
}
\description{
Implementation of W3.CSS Buttons, as described in
\url{https://www.w3schools.com/w3css/w3css_buttons.asp}
}
\examples{
w3_actionButton("ok", "Go")

}
