% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3css_badges.R
\name{w3_badge}
\alias{w3_badge}
\title{W3.CSS Badges}
\usage{
w3_badge()
}
\value{
A CSS class.
}
\description{
Implementation of W3.CSS Badges, as described in
\url{https://www.w3schools.com/w3css/w3css_badges.asp}
}
\examples{
w3_badge()
}
