% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{w3_colors}
\alias{w3_colors}
\title{W3.CSS colors}
\format{
A vector of colors
}
\source{
\url{https://www.w3schools.com/w3css/w3css_colors.asp}
}
\usage{
w3_colors
}
\description{
W3.CSS colors
}
\keyword{datasets}
