% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3css_containers.R
\name{w3_container}
\alias{w3_container}
\title{W3.CSS Containers}
\usage{
w3_container(..., type = "div")
}
\arguments{
\item{...}{Content of the container}

\item{type}{HTML tag to use as a container. Default is \code{div}.}
}
\value{
A tag.
}
\description{
Implementation of W3.CSS Containers, as described in
\url{https://www.w3schools.com/w3css/w3css_containers.asp}
}
\examples{
# Simple container
w3_container("hey there")
# Container with a 'teal'-colored background
w3_container("hey there", class = w3_color("teal"))
}
