% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3css_dropdowns.R
\name{w3_dropdown}
\alias{w3_dropdown}
\title{W3.CSS Dropdowns}
\usage{
w3_dropdown(
  button_text = "Hover Over Me!",
  content = list(),
  button_class = NULL,
  content_style = NULL
)
}
\arguments{
\item{button_text}{Text to display inside the button}

\item{content}{Content of the dropdown}

\item{button_class}{Additional class to add to the button}

\item{content_style}{Additional style to add to the content tag}
}
\value{

}
\description{
Implementation of W3.CSS Dropdowns, as described in
\url{https://www.w3schools.com/w3css/w3css_dropdowns.asp}
}
\examples{

}
