% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3css_padding.R
\name{w3_padding}
\alias{w3_padding}
\title{W3.CSS Padding}
\usage{
w3_padding(value = c("", 16, 24, 32, 48, 64, "small", "large"))
}
\arguments{
\item{value}{The value of the padding. Must be one of  "", 16, 24, 32, 48,
64, "small", or "large".}
}
\value{
A CSS class
}
\description{
Implementation of W3.CSS Padding, as described in
\url{https://www.w3schools.com/w3css/w3css_padding.asp}
}
\examples{
w3_padding(16)
}
