% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{w3_page}
\alias{w3_page}
\title{W3 CSS Page}
\usage{
w3_page(...)
}
\arguments{
\item{...}{Content of the page}
}
\value{
A shiny UI using W3.CSS
}
\description{
W3 CSS Page
}
\examples{
w3_page("hey there")
}
