% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3css_panels.R
\name{w3_panel}
\alias{w3_panel}
\title{W3.CSS Panels}
\usage{
w3_panel(..., with_close = FALSE)
}
\arguments{
\item{...}{Content of the panel.}

\item{with_close}{Add a close button to the panel.}
}
\value{
A tag.
}
\description{
Implementation of W3.CSS colors, as described in
\href{https://www.w3schools.com/w3css/w3css_panels.asp}{www.w3schools.com/w3css/w3css_panels.asp}.
}
\examples{
# Simple container
w3_panel("hey there")
# Container with a 'teal'-colored background
w3_panel("hey there", class = w3_color("teal"))
}
