% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3css_borders.R
\name{w3_border}
\alias{w3_border}
\alias{w3_border_top}
\alias{w3_border_right}
\alias{w3_border_bottom}
\alias{w3_border_left}
\alias{w3_border_0}
\alias{w3_border_color}
\alias{w3_hover_border_color}
\alias{w3_bottombar}
\alias{w3_leftbar}
\alias{w3_rightbar}
\alias{w3_topbar}
\title{W3.CSS Borders}
\usage{
w3_border()

w3_border_top()

w3_border_right()

w3_border_bottom()

w3_border_left()

w3_border_0()

w3_border_color(color)

w3_hover_border_color(color)

w3_bottombar()

w3_leftbar(color)

w3_rightbar(color)

w3_topbar(color)
}
\arguments{
\item{color}{One of W3 css color}
}
\value{
A W3 CSS class.
}
\description{
Implementation of W3.CSS borders, as described in
\href{https://www.w3schools.com/w3css/w3css_borders.asp}{www.w3schools.com/w3css/w3css_borders.asp}.
}
\examples{
w3_border()
w3_border_top()
}
