% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3css_colors.R
\name{w3_color}
\alias{w3_color}
\alias{w3_text_color}
\alias{w3_hover_color}
\title{W3.CSS Colors}
\usage{
w3_color(color)

w3_text_color(color)

w3_hover_color(color)
}
\arguments{
\item{color}{One of W3 css color}
}
\value{
A css class
}
\description{
Implementation of W3.CSS colors, as described in
\href{https://www.w3schools.com/w3css/w3css_colors.asp}{www.w3schools.com/w3css/w3css_colors.asp}.
}
\examples{
w3_color("red")
w3_color("pink")
w3_color("blue")
}
